+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.10.11
+* Fixed an issue when commit dialog won't show on a large number of files.
 * Fixed Bash button when working with UNC paths on Windows by [Fez Vrasta](https://github.com/FezVrasta)
 
 ## 0.10.10